created new mock file for dataWeather
