+++ src/mocks/weather.ts
@@ -1,0 +1,48 @@
+export const listTab = [
+    {
+        status: 'Today',
+    },
+    {
+        status: 'Tomorrow',
+    },
+    {
+        status: 'Next 7 Days',
+    },
+];
+
+export const data = [
+    {
+
+        status: 'Today',
+    },
+    {
+        status: 'Today',
+    },
+    {
+        status: 'Today',
+    },
+    {
+        status: 'Today',
+    },
+    {
+        status: 'Today',
+    },
+    {
+        status: 'Today',
+    },
+    {
+        status: 'Today',
+    },
+    {
+        status: 'Tomorrow',
+    },
+    {
+        status: 'Tomorrow',
+    },
+    {
+        status: 'Tomorrow',
+    },
+    {
+        status: 'Next 7 Days',
+    },
+];
