--- src/request/openWeatherAPI/dataWeather.js
@@ -1,50 +1,0 @@
-// data Tabs
-
-export const listTab = [
-  {
-    status: 'Today',
-  },
-  {
-    status: 'Tomorrow',
-  },
-  {
-    status: 'Next 7 Days',
-  },
-];
-
-export const data = [
-  {
-
-    status: 'Today',
-  },
-  {
-    status: 'Today',
-  },
-  {
-    status: 'Today',
-  },
-  {
-    status: 'Today',
-  },
-  {
-    status: 'Today',
-  },
-  {
-    status: 'Today',
-  },
-  {
-    status: 'Today',
-  },
-  {
-    status: 'Tomorrow',
-  },
-  {
-    status: 'Tomorrow',
-  },
-  {
-    status: 'Tomorrow',
-  },
-  {
-    status: 'Next 7 Days',
-  },
-];
